+++ example/ios/RunnerTests/RunnerTests.swift
@@ -3,7 +3,7 @@ import UIKit
 import XCTest
 
 
-@testable import upi_payment_plugin
+@testable import deep_upi
 
 // This demonstrates a simple unit test of the Swift portion of this plugin's implementation.
 //